Refactor Spinner and LoadMore
